move  to edge.ts
