+++ packages/uploadthing/server.ts
@@ -2,6 +2,6 @@ import { createBuilder } from "./src/upload-builder";
 
 export * from "./src/types";
 export * as utapi from "./src/helper-functions";
-export { createServerHandler } from "./src/next/core/approuter";
+export { createServerHandler } from "./src/internal/edge";
 
 export const createUploadthing = () => createBuilder<"web">();
