+++ packages/uploadthing/src/next/core/approuter.ts
@@ -53,16 +53,3 @@ export const createNextRouteHandler = <TRouter extends FileRouter>(
 
   return { GET, POST };
 };
-
-export const createServerHandler = (
-  ...args: Parameters<typeof createNextRouteHandler>
-) => {
-  const handler = createNextRouteHandler(...args);
-  const thisHandler =
-    (method: keyof typeof handler) => (event: { request: Request } | Request) =>
-      handler[method](event instanceof Request ? event : event.request);
-  return {
-    GET: thisHandler("GET"),
-    POST: thisHandler("POST"),
-  };
-};
